- Funkce které byly původně přesunuty z ComplexMat do KCF_Tracker::Model nyní přesunuty do nového souboru cvmat_func.h
